Moving new classes into subclasses of the old ones
